* Fix missing curve reference in ECC key generation * Add BLS_ECDH_X support for secp256r1 * Clean up enclave memory more aggressively * Move to production keys for script validation and IAS * Replace assert by platform_assert to avoid Unlimited bugs (tm) * Fix missing include in bolos.h, few casts and unused variables * Add Keccak test * Move to production enclave configuration * Update dependencies patches for Windows build * Add SIMU flag for enclave builds to support the simulator key set * Add release images and usage guide * Add Docker files for Linux deterministic builds * Add Windows build * Add host client

+++ tee/moxie/moxie_swi_sodium.c
@@ -498,13 +498,13 @@ void moxie_crypto_sign(struct machine *mach) {
     }
 
     status = crypto_sign(sm, &smlen_param, m, mlen, sk);
-    if (!write32_check(mach, mach->cpu.regs[MOXIE_R1], smlen_param)) {
+    if (!write32_check(mach, mach->cpu.regs[MOXIE_R1], (uint32_t)smlen_param)) {
         DEBUG("Invalid length pointer (write)\n");
         status = 0;
         goto end;
     }
     {
-        uint32_t test = smlen_param;
+        uint32_t test = (uint32_t)smlen_param;
         DEBUG("Sign writing size %d %d\n", test, crypto_sign_BYTES);
     }
 
@@ -550,7 +550,7 @@ void moxie_crypto_sign_open(struct machine *mach) {
     }
 
     status = crypto_sign_open(m, &mlen_param, sm, smlen, pk);
-    if (!write32_check(mach, mach->cpu.regs[MOXIE_R1], mlen_param)) {
+    if (!write32_check(mach, mach->cpu.regs[MOXIE_R1], (uint32_t)mlen_param)) {
         DEBUG("Invalid length pointer (write)\n");
         status = 0;
         goto end;
